pyproject.toml
[project]
name = "ollamaparser"
version = "0.1.0"
description = "ollamaparser: PDF -> layout detection (DocLayout-YOLO/PDFium) + OCR (DeepSeek-OCR via Ollama) + VLM (Gemma3 via Ollama) -> Markdown + metadata"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pypdfium2==4.30.0",
  "pillow==11.0.0",
  "typer==0.15.1",
  "click==8.1.7",
  "pydantic==2.10.3",
  "rich==13.9.4",
  "python-dotenv==1.0.1",
  "scikit-learn==1.5.2",
  "numpy==2.1.3",
  "requests==2.32.3",
]

[project.optional-dependencies]
detector = [
  # DocLayout-YOLO official SDK (YOLOv10). See upstream quickstart.
  "doclayout-yolo",
  "huggingface-hub",
]

[project.scripts]
ollamaparser-parse = "ollamaparser.cli:main"

[tool.uv]
# Install this project into the environment (so `project.scripts` entry points work).
package = true

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"


